soc/integration: add csr ordering support
csr_ordering default is "big" but there was no way for software
to detect if csr_ordering was changed to "little" to match the cpu.

tools/litex_json2dts_linux: add *-endian to support csr ordering

csr_ordering default is "big" but there was no way for software
to detect if csr_ordering was changed to "little" to match the cpu.